+++ src/animatedclusters_ui.cpp
@@ -191,8 +191,14 @@ void AnimatedClusters::processUI(double time, nvh::Profiler& profiler, const Cal
                  "uses the nv_cluster_builder library, otherwise meshoptimizer");
     PE::InputFloat("NV graph weight", &m_sceneConfig.clusterNvGraphWeight, 0.01f, 0.01f, "%.3f", ImGuiInputTextFlags_EnterReturnsTrue,
                    "non-zero weight makes use of triangle connectivity, otherwise disabled");
+    PE::InputFloat("NV cost underfill", &m_sceneConfig.clusterNvUnderfill, 0.01f, 0.01f, "%.3f",
+                   ImGuiInputTextFlags_EnterReturnsTrue, "cost to underfilling a cluster");
+    PE::InputFloat("NV cost overlap", &m_sceneConfig.clusterNvOverlap, 0.01f, 0.01f, "%.3f",
+                   ImGuiInputTextFlags_EnterReturnsTrue, "cost to bounding box overlap between clusters");
     PE::Checkbox("Optimize for triangle strips", &m_sceneConfig.clusterStripify,
                  "Re-order triangles within cluster optimizing for triangle strips");
+    PE::Checkbox("Cluster-dedicated vertices", &m_sceneConfig.clusterDedicatedVertices,
+                 "stores vertices per cluster, increases memory / animation work");
     PE::end();
 
     ImGui::TextDisabled("Ray tracer CLAS specifics");